Use PHPMailer to send 'raw' emails.
_og_mailinglist_send_raw_email has been using the system command to
directly invoke sendmail with the message we want to send. This patch
switches us to using PHPMailer. There will be some minor differences
in the final result - we cannot easily override PHPMailer's setting
of the message date to make it exactly the same as the incoming
message. It seems to me to be close enough, though.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>