Add support for the Drupal PHPMailer module.
Add support for the Drupal PHPMailer module, which provides
the user with an easy way of setting SMTP servers and so on,
without needing to edit the source itself. If this module
is unavailable, we fall back to the PHPMailer library.

Note that there is a slight difference in behaviour: the
Drupal PHPMailer module uses SMTP only, while the library
itself defaults to using sendmail support).

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>