Check that parent_message_id isn't empty.
Should fix bug were new posts via email are added to old posts as comments. The problem was that the
parent_message_id that was returned was empty but that was still used in
og_mailinglist_get_nid_from_parent_messageid() to query for a
parent_message_id. The nid returned was from a message that didn't have
a message-id in the database for whatever reason.